add env.close
when you run many envs, you have to have a way to close the render windows. env.close() is the standard way to do this in openai gym, so this pull request just adds a simple env.close method to the TradingEnv class, which closes the pyplot window. then you can play atari games and trade stocks in a loop with rendering, and not clutter your screen with too many windows!